+++ ChangeLog.md
@@ -9,6 +9,7 @@
     - Shell functions trap_int() and trap_exit() output to stderr
     - Deprecate 'Environment' property in favor of 'environment' (todo #65), thus bumping minor candidate version number
     - Deprecate 'environment.type' property in favor of 'environment.id' (todo #66)
+    - Deprecate 'MQTTGateway.broker' property in favor of 'MQTTGateway.host' (todo #62)
 
 ### 4.1.3